refactor: 외부 AI 서버 연결 API, ChatMessage 관련 코드 리팩토링
