+++ src/main/java/com/lawProject/SSL/domain/langchain/api/OllamaApiController.java
@@ -3,7 +3,6 @@ package com.lawProject.SSL.domain.langchain.api;
 import com.lawProject.SSL.domain.langchain.dto.MakeDocForm;
 import com.lawProject.SSL.domain.langchain.service.OllamaApiClient;
 import com.lawProject.SSL.domain.lawsuit.dto.FileStorageResult;
-import com.lawProject.SSL.domain.lawsuit.service.LawSuitService;
 import com.lawProject.SSL.global.common.code.ErrorCode;
 import com.lawProject.SSL.global.common.code.SuccessCode;
 import com.lawProject.SSL.global.common.response.ApiResponse;
@@ -28,7 +27,6 @@ import java.time.LocalDateTime;
 @RequestMapping("/api/v1/ollama")
 public class OllamaApiController {
     private final OllamaApiClient ollamaApiClient;
-    private final LawSuitService lawSuitService;
     @Resource(name = "chatRedisTemplate")
     private final RedisTemplate<String, Object> redisTemplate;
     private final ChannelTopic topic;
