+++ src/main/java/com/lawProject/SSL/domain/langchain/repository/ChatMessageRepository.java
@@ -1,4 +1,4 @@
-package com.lawProject.SSL.domain.langchain.dao;
+package com.lawProject.SSL.domain.langchain.repository;
 
 import com.lawProject.SSL.domain.langchain.domain.ChatMessage;
 import com.lawProject.SSL.domain.chatroom.model.ChatRoom;
@@ -6,7 +6,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MessageRepository extends JpaRepository<ChatMessage, Long> {
+public interface ChatMessageRepository extends JpaRepository<ChatMessage, Long> {
 
     Page<ChatMessage> findByChatRoom(Pageable pageable, ChatRoom chatRoom);
